+++ app/Http/Controllers/Api/V1/CostCenterController.php
@@ -9,36 +9,81 @@ use App\Http\Requests\StoreCostCenterRequest;
 use App\Http\Resources\Finance\CostCenterResource;
 use App\Models\Tenant\CostCenter;
 use Illuminate\Http\Response;
+use Spatie\Permission\Exceptions\PermissionDoesNotExist;
+use Illuminate\Http\Request;
 
 class CostCenterController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
-        $costCenters = CostCenter::paginate(10);
-        return CostCenterResource::collection($costCenters);
+        try {
+            $user = $request->user();
+            if ($user->hasPermissionTo('cost_center_index')) {
+                return CostCenterResource::collection(CostCenter::paginate(20));
+            } else {
+                return response()->json(['message' => 'Unauthorized for this task'], 403);
+            }
+        } catch (PermissionDoesNotExist $exception) {
+            return response()->json(['message' => 'Unauthorized for this task - no permission by this name'], 403);
+        }
     }
-
 
     public function store(StoreCostCenterRequest $request)
     {
-        $costCenter = CostCenter::create($request->validated());
-        return response()->json(['data' => new CostCenterResource($costCenter)], Response::HTTP_CREATED);
+        try {
+            $user = $request->user();
+            if ($user->hasPermissionTo('cost_center_store')) {
+                $costCenter = CostCenter::create($request->validated());
+                return new CostCenterResource($costCenter);
+            } else {
+                return response()->json(['message' => 'Unauthorized for this task'], 403);
+            }
+        } catch (PermissionDoesNotExist $exception) {
+            return response()->json(['message' => 'Unauthorized for this task - no permission by this name'], 403);
+        }
     }
 
-    public function show(CostCenter $costCenter)
+    public function show(Request $request, CostCenter $costCenter)
     {
-        return response()->json(['data' => new CostCenterResource($costCenter)], Response::HTTP_OK);
+        try {
+            $user = $request->user();
+            if ($user->hasPermissionTo('cost_center_show')) {
+                return new CostCenterResource($costCenter);
+            } else {
+                return response()->json(['message' => 'Unauthorized for this task'], 403);
+            }
+        } catch (PermissionDoesNotExist $exception) {
+            return response()->json(['message' => 'Unauthorized for this task - no permission by this name'], 403);
+        }
     }
 
     public function update(UpdateCostCenterRequest $request, CostCenter $costCenter)
     {
-        $costCenter->update($request->validated());
-        return response()->json(['data' => new CostCenterResource($costCenter)], Response::HTTP_OK);
+        try {
+            $user = $request->user();
+            if ($user->hasPermissionTo('cost_center_update')) {
+                $costCenter->update($request->validated());
+                return new CostCenterResource($costCenter);
+            } else {
+                return response()->json(['message' => 'Unauthorized for this task'], 403);
+            }
+        } catch (PermissionDoesNotExist $exception) {
+            return response()->json(['message' => 'Unauthorized for this task - no permission by this name'], 403);
+        }
     }
 
-    public function destroy(CostCenter $costCenter)
+    public function destroy(Request $request, CostCenter $costCenter)
     {
-        $costCenter->delete();
-        return response()->json(null, Response::HTTP_NO_CONTENT);
+        try {
+            $user = $request->user();
+            if ($user->hasPermissionTo('cost_center_destroy')) {
+                $costCenter->delete();
+                return response()->json(['message' => 'Cost center deleted successfully'], 200);
+            } else {
+                return response()->json(['message' => 'Unauthorized for this task'], 403);
+            }
+        } catch (PermissionDoesNotExist $exception) {
+            return response()->json(['message' => 'Unauthorized for this task - no permission by this name'], 403);
+        }
     }
 }
